Version 1.2
This Update Adds a Working Heart system.
(Release Notes at the bottom!)
You take damage and the heart(Shield) starts to break.
Once the shield is completely broken, You will die once you hit a asteroid and lose a life.
Don't worry, Your points will not be Lost.
I also made it so that you need more points to beat the game as compensation for the game isn't too easy.
-Added health System.
-Increase amount of points to win